MetalImageFilters: Version 1.0, 2016-09-13
First release

This sample demonstrates how to use a selection of MPSKernel classes to filter an image on the GPU. The image is processed as a MTLTexture object, obtained from either a bundled file or a real-time video stream. The filters demonstrated range from a single-pass Sobel filter to a multi-pass histogram specification.