IceCreamBuilder: Version 1.0, 2016-09-13
First release

This sample is a simple example of building an app extension that interacts with the Messages app and lets users send interactive messages and create stickers.

The extension is based on a MSMessagesAppViewController subclass that then presents a child view controller depending on the current conversation, message state and presentation style. It also updates a conversation with new or updated messages and posts stickers to a conversation.