User Interface 3D Transforms: Version 1.0, 2017-04-27
First version.

This sample code demonstrates applying 3D effects to an arbitrary UIView. A few different views are chosen randomly to facilitate this example: an image view, map view, and a web view. A view hierarchy is displayed that is akin to the tab switching user experience in Safari on iOS and the 3D nature of this effect is implemented using CALayer's CATransform3D.

Signed-off-by: Liu Lantao <liulantao@gmail.com>