PageControl: Version 7.0, 2017-09-21
Upgraded to Swift, now uses Storyboards.

This application demonstrates how to use a "paginated" UIScrollView to display pages of images in a horizontal fashion as a mechanism for navigating between different pages of content. It is optimized to minimize the memory usage by loading only three image pages at a time the user user scrolls: the current page, the page before and after it. This application was designed as universal to run on both iPhone and iPad devices.

Signed-off-by: Liu Lantao <liulantao@gmail.com>