ApplicationShortcuts: Version 1.2, 2016-09-28
Upgraded to Swift 2.3.

Demonstrates how to use the UIApplicationShortcutItem class to provide quick access to parts of your application directly from the device's home screen. The sample shows two static shortcuts (defined in the app's Info.plist), and two dynamic shortcuts (defined in code with the UIMutableApplicationShortcutItem class). The dynamic shortcuts can be edited to change the title, subtitle and icon.

Signed-off-by: Liu Lantao <liulantao@gmail.com>