AudioInARKit - Audio in ARKit: Version 1.1, 2017-09-19
Editorial update.

This sample app runs an ARKit world tracking session with content displayed in a SceneKit view. To demonstrate plane detection, the app simply places a 3D model onto the first plane that ARKit detects. If the model's position is outside the current field of view of the camera, the app uses SceneKit's positional audio feature to indicate which direction to turn the device to see the model.

Signed-off-by: Liu Lantao <liulantao@gmail.com>