PhotoPicker: Using UIImagePickerController to Select Pictures and Take Photos: Version 4.2, 2018-01-11
Updated to iOS 11 SDK, converted to Swift, storyboard changes to support iPhone X.

PhotoPicker demonstrates choosing images from the photo library, take a picture with the camera, and customizing the camera view. This is done by using UIImagePickerController. The chosen image or camera photo is displayed in a UIImageView.

To customize the camera's interface, this sample shows how to use an overlay view. With this overlay view it gives you the ability to customize the UI as you take a picture.

Among the custom features of the camera is to take a single picture, timed picture, or repeated pictures like a camera with a fast shutter speed. Timed and shutter speed camera shots are done using the NSTimer class.

Signed-off-by: Liu Lantao <liulantao@gmail.com>