CloudCaptions: How integrate CloudKit into your application: Version 1.0, 2014-09-17
First public release.

This sample shows how to use CloudKit to upload and retrieve CKRecords and associated assets. In this example, there are two record types, an image record type and a post record type. Users are able to upload their own photos or select an image already found in an image record type. This example also uses an NSPredicate in its CKQueries to filter results based on tags.

Signed-off-by: Liu Lantao <liulantao@gmail.com>